+++ analyzer/src/main/java/nl/basjes/parse/useragent/parse/EvilManualUseragentStringHacks.java
@@ -133,7 +133,10 @@ public final class EvilManualUseragentStringHacks {
         // The Weibo useragent This one is a single useragent that hold significant traffic
         result = replaceString(result, "__", " ");
 
-        if (result.contains("%20") || result.contains("%3B")) {
+        if (result.contains("%20") ||
+            result.contains("%3B") ||
+            result.contains("%25") ||
+            result.contains("%2F")) {
             try {
                 result = URLDecoder.decode(result, "UTF-8");
             } catch (UnsupportedEncodingException | IllegalArgumentException e) {