+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v5.15-SNAPSHOT
   - Better url extraction.
   - Fix NPE when defining invalid rules (walk left/right/up from the top of the tree)
   - The Tencent/Alibaba NetType and Language tags are now better extracted.
+  - AgentName is now more consistent (konqueror/Konqueror)
 - New/improved detections
   - App: Desktop apps using Electron (https://electronjs.org/)
   - Devices: Improve detection of Apple devices