+++ CHANGELOG.md
@@ -16,6 +16,7 @@ v5.15-SNAPSHOT
   - Apache Beam   2.17.0
 - Build
   - Improve scripting for generated files
+  - Only build the docker image for the webservlet if docker is available.
 - Demonstration WebServlet
   - Updated to SpringFox 3.0.0-SNAPSHOT to support better examples in the Swagger UI.